Move useFloatingDropdown to hooks directory
